Download and import an empty FDZ file.

Download and import an empty data set from the FDZ homepage.

Usage

download_fdz(
  study = c("PISA"),
  year = c("2018", "2015", "2012", "2009", "2006", "2003", "2000"),
  data_type = c("stud_quest")
)

Arguments

study

Name of the study

year

Year of the assessment (only needed for longitudinal assessments).

data_type

Type of the data.

Details

The function downloads and imports an empty data set (Leerdatensatz) from the FDZ homepage. These data sets contain zero rows. The data is imported via import_spss as a GADSdat object.

Examples

fdz_pisa <- download_fdz(study = "PISA", year = "2015",
                        data_type = "stud_quest")